Store impact in CVSS classes
Part of #683

+++ src/main/java/com/sap/oss/phosphor/fosstars/data/owasp/model/Cvssv2.java
@@ -13,7 +13,7 @@ public class Cvssv2 implements Serializable {
   private float score;
 
   @JsonProperty("score")
-  public float getScore() {
+  public double getScore() {
     return score;
   }
 }
